test(mf2): Split test data into three files; leave out coverage from bare `npm test`
