+++ .github/workflows/nodejs.yml
@@ -18,7 +18,7 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
       - run: npm ci
-      - run: npm test
+      - run: npm test -- --coverage
       - run: npm run build
 
   lint:
