+++ packages/mf2-messageformat/src/__fixtures/syntax-errors.json
@@ -1,0 +1,58 @@
+[
+  ".",
+  "{",
+  "}",
+  "{}",
+  "{{",
+  "{{}",
+  "{{}}}",
+  "{|foo| #markup}",
+  "{/tag foo=bar}",
+  "{{missing end brace}",
+  "{{missing end braces",
+  "{{missing end {$braces",
+  "{{extra}} content",
+  "empty { } placeholder",
+  "bad {:} placeholder",
+  "bad {\\u0000placeholder}",
+  "no-equal {|42| :number minimumFractionDigits 2}",
+  "bad {:placeholder option=x option=x}",
+  "bad {:placeholder ns:option=x ns:option=y}",
+  "bad {:placeholder option=}",
+  "bad {:placeholder option value}",
+  "bad {:placeholder option:value}",
+  "bad {:placeholder option}",
+  "bad {:placeholder:}",
+  "bad {::placeholder}",
+  "bad {:placeholder::foo}",
+  "bad {:placeholder option:=x}",
+  "bad {:placeholder :option=x}",
+  "bad {:placeholder option::x=y}",
+  "bad {$placeholder option}",
+  "no {placeholder end",
+  "no {$placeholder end",
+  "no {:placeholder end",
+  "no {|placeholder| end",
+  "no {|literal} end",
+  "no {|literal or placeholder end",
+  ".input {$foo} .local $foo = {42} {{bar {$foo}}}",
+  ".local $foo = {42} .input {$foo} {{bar {$foo}}}",
+  ".local $foo = {:unknown} .local $foo = {42} {{bar {$foo}}}",
+  ".local bar = {|foo|} {{{$bar}}}",
+  ".local $bar {|foo|} {{{$bar}}}",
+  ".local $bar = |foo| {{{$bar}}}",
+  ".local $foo = {$bar} .local $bar = {$baz} {{bar {$foo}}}",
+  ".local $foo = {$bar :func} .local $bar = {$baz} {{bar {$foo}}}",
+  ".match {#foo} * {{foo}}",
+  ".match {} * {{foo}}",
+  ".match {|foo| :x} {|bar| :x} ** {{foo}}",
+  ".match * {{foo}}",
+  ".match {|x| :x} * foo",
+  ".match {|x| :x} * {{foo}} extra",
+  ".match |x| * {{foo}}",
+  ".match {$foo :x} * * {{foo}}",
+  ".match {$foo :x} {$bar :x} * {{foo}}",
+  ".match {$foo} one {{one}} * {{other}}",
+  ".input {$foo} .match {$foo} one {{one}} * {{other}}",
+  ".local $foo = {$bar} .match {$foo} one {{one}} * {{other}}"
+]
