+++ packages/mf2-messageformat/src/__fixtures/test-functions.json
@@ -1,0 +1,187 @@
+{
+  "number": [
+    { "src": "hello {4.2 :number}", "exp": "hello 4.2" },
+    { "src": "hello {-4.20 :number}", "exp": "hello -4.2" },
+    { "src": "hello {0.42e+1 :number}", "exp": "hello 4.2" },
+    {
+      "src": "hello {foo :number}",
+      "exp": "hello {|foo|}",
+      "errors": [{ "type": "bad-input" }]
+    },
+    {
+      "src": "hello {:number}",
+      "exp": "hello {:number}",
+      "errors": [{ "type": "bad-input" }]
+    },
+    {
+      "src": "hello {4.2 :number minimumFractionDigits=2}",
+      "exp": "hello 4.20"
+    },
+    {
+      "src": "hello {|4.2| :number minimumFractionDigits=|2|}",
+      "exp": "hello 4.20"
+    },
+    {
+      "src": "hello {4.2 :number minimumFractionDigits=$foo}",
+      "params": { "foo": 2 },
+      "exp": "hello 4.20"
+    },
+    {
+      "src": "hello {|4.2| :number minimumFractionDigits=$foo}",
+      "params": { "foo": "2" },
+      "exp": "hello 4.20"
+    },
+    {
+      "src": ".local $foo = {$bar :number} {{bar {$foo}}}",
+      "params": { "bar": 4.2 },
+      "exp": "bar 4.2"
+    },
+    {
+      "src": ".local $foo = {$bar :number minimumFractionDigits=2} {{bar {$foo}}}",
+      "params": { "bar": 4.2 },
+      "exp": "bar 4.20"
+    },
+    {
+      "src": ".local $foo = {$bar :number minimumFractionDigits=foo} {{bar {$foo}}}",
+      "params": { "bar": 4.2 },
+      "exp": "bar {$bar}",
+      "errors": [{ "type": "bad-option" }]
+    },
+    {
+      "src": ".local $foo = {$bar :number} {{bar {$foo}}}",
+      "params": { "bar": "foo" },
+      "exp": "bar {$bar}",
+      "errors": [{ "type": "bad-input" }]
+    },
+    {
+      "src": ".input {$foo :number} {{bar {$foo}}}",
+      "params": { "foo": 4.2 },
+      "exp": "bar 4.2"
+    },
+    {
+      "src": ".input {$foo :number minimumFractionDigits=2} {{bar {$foo}}}",
+      "params": { "foo": 4.2 },
+      "exp": "bar 4.20"
+    },
+    {
+      "src": ".input {$foo :number minimumFractionDigits=foo} {{bar {$foo}}}",
+      "params": { "foo": 4.2 },
+      "exp": "bar {$foo}",
+      "errors": [{ "type": "bad-option" }]
+    },
+    {
+      "src": ".input {$foo :number} {{bar {$foo}}}",
+      "params": { "foo": "foo" },
+      "exp": "bar {$foo}",
+      "errors": [{ "type": "bad-input" }]
+    },
+    {
+      "src": ".match {$foo :number} one {{one}} * {{other}}",
+      "params": { "foo": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".match {$foo :number} 1 {{=1}} one {{one}} * {{other}}",
+      "params": { "foo": 1 },
+      "exp": "=1"
+    },
+    {
+      "src": ".match {$foo :number} one {{one}} 1 {{=1}} * {{other}}",
+      "params": { "foo": 1 },
+      "exp": "=1"
+    },
+    {
+      "src": ".match {$foo :number} {$bar :number} one one {{one one}} one * {{one other}} * * {{other}}",
+      "params": { "foo": 1, "bar": 1 },
+      "exp": "one one"
+    },
+    {
+      "src": ".match {$foo :number} {$bar :number} one one {{one one}} one * {{one other}} * * {{other}}",
+      "params": { "foo": 1, "bar": 2 },
+      "exp": "one other"
+    },
+    {
+      "src": ".match {$foo :number} {$bar :number} one one {{one one}} one * {{one other}} * * {{other}}",
+      "params": { "foo": 2, "bar": 2 },
+      "exp": "other"
+    },
+    {
+      "src": ".input {$foo :number} .match {$foo} one {{one}} * {{other}}",
+      "params": { "foo": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".local $foo = {$bar :number} .match {$foo} one {{one}} * {{other}}",
+      "params": { "bar": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".input {$foo :number } .local $bar = {$foo} .match {$bar} one {{one}} * {{other}}",
+      "params": { "foo": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".input {$bar :number} .match {$bar} one {{one}} * {{other}}",
+      "params": { "bar": 2 },
+      "exp": "other"
+    },
+    {
+      "src": ".input {$bar} .match {$bar :number} one {{one}} * {{other}}",
+      "params": { "bar": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".input {$bar} .match {$bar :number} one {{one}} * {{other}}",
+      "params": { "bar": 2 },
+      "exp": "other"
+    },
+    {
+      "src": ".input {$bar} .match {$bar :number} one {{one}} * {{other}}",
+      "params": { "bar": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".input {$bar} .match {$bar :number} one {{one}} * {{other}}",
+      "params": { "bar": 2 },
+      "exp": "other"
+    },
+    {
+      "src": ".input {$none} .match {$foo :number} one {{one}} * {{{$none}}}",
+      "params": { "foo": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".local $bar = {$none} .match {$foo :number} one {{one}} * {{{$bar}}}",
+      "params": { "foo": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".local $bar = {$none} .match {$foo :number} one {{one}} * {{{$bar}}}",
+      "params": { "foo": 2 },
+      "exp": "{$none}",
+      "errors": [{ "type": "unresolved-var" }]
+    }
+  ],
+  "string": [
+    {
+      "src": ".match {$foo :string} |1| {{one}} * {{other}}",
+      "params": { "foo": "1" },
+      "exp": "one"
+    },
+    {
+      "src": ".match {$foo :string} 1 {{one}} * {{other}}",
+      "params": { "foo": 1 },
+      "exp": "one"
+    },
+    {
+      "src": ".match {$foo :string} 1 {{one}} * {{other}}",
+      "params": { "foo": null },
+      "exp": "other"
+    },
+    {
+      "src": ".match {$foo :string} 1 {{one}} * {{other}}",
+      "exp": "other",
+      "errors": [{ "type": "unresolved-var" }]
+    }
+  ]
+}
