+++ packages/mf2-messageformat/src/messageformat.test.ts
@@ -1,30 +1,34 @@
-import testMessages from './__fixtures/test-messages.json';
-import { testName } from './__fixtures/test-name';
+import syntaxErrors from './__fixtures/syntax-errors.json';
+import testCore from './__fixtures/test-core.json';
+import testFunctions from './__fixtures/test-functions.json';
+import { testName } from './__fixtures/util-test-name';
 import { MessageSyntaxError } from './errors';
 import { MessageFormat } from './messageformat';
 
-export type TestMessage =
-  | {
-      src: string;
-      only?: boolean;
-      locale?: string;
-      params?: Record<string, string | number | null | undefined>;
-      exp: string;
-      parts?: Array<object>;
-      errors?: Array<{ type: string }>;
-      syntaxError?: never;
-    }
-  | { src: string; syntaxError: true };
+export type TestMessage = {
+  src: string;
+  only?: boolean;
+  locale?: string;
+  params?: Record<string, string | number | null | undefined>;
+  exp: string;
+  parts?: Array<object>;
+  errors?: Array<{ type: string }>;
+};
 
-describe('Format messages', () => {
-  for (const testMsg of testMessages as TestMessage[]) {
-    if (testMsg.syntaxError) {
-      test(testName(testMsg), () => {
-        expect(() => new MessageFormat(testMsg.src)).toThrow(
-          MessageSyntaxError
-        );
-      });
-    } else {
+describe('Syntax errors', () => {
+  for (const src of syntaxErrors) {
+    test(src, () => {
+      expect(() => new MessageFormat(src)).toThrow(MessageSyntaxError);
+    });
+  }
+});
+
+for (const [title, messages] of [
+  ['Core message syntax', testCore],
+  ...Object.entries(testFunctions).map(x => [`Messages using :${x[0]}`, x[1]])
+] as Array<[string, TestMessage[]]>) {
+  describe(title, () => {
+    for (const testMsg of messages) {
       const {
         src,
         exp,
@@ -48,5 +52,5 @@ describe('Format messages', () => {
         }
       });
     }
-  }
-});
+  });
+}
